+++ common/src/main/java/dev/terminalmc/clientsort/mixin/client/CreativeModeInventoryScreenMixin.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2025 TerminalMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.terminalmc.clientsort.mixin.client;
+
+import dev.terminalmc.clientsort.client.gui.TriggerButtonManager;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.CreativeModeTab.Type;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(CreativeModeInventoryScreen.class)
+public class CreativeModeInventoryScreenMixin extends Screen {
+
+    protected CreativeModeInventoryScreenMixin(Component title) {
+        super(title);
+    }
+
+    @Inject(
+            method = "selectTab",
+            at = @At("RETURN")
+    )
+    private void afterSelectTab(CreativeModeTab tab, CallbackInfo ci) {
+        TriggerButtonManager.getPlayerButtons().forEach(this::removeWidget);
+        if (tab.getType().equals(Type.INVENTORY))
+            TriggerButtonManager.afterScreenInit(this);
+    }
+}