+++ fabric/src/main/java/dev/terminalmc/clientsort/client/ClientSortFabric.java
@@ -22,7 +22,6 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
-import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
 import net.fabricmc.loader.api.FabricLoader;
@@ -40,10 +39,6 @@ public class ClientSortFabric implements ClientModInitializer {
 
         // Register after-tick event
         ClientTickEvents.END_CLIENT_TICK.register(ClientSort::afterClientTick);
-
-        // Register screen after-init event
-        ScreenEvents.AFTER_INIT.register(
-                (mc, screen, scaledWidth, scaledHeight) -> ClientSort.afterScreenInit(screen));
 
         // Register all custom S2C payload handlers
         ClientRegistration.PAYLOADS_S2C.forEach(ClientSortFabric::registerHandlerS2C);