+++ changelog.md
@@ -4,6 +4,8 @@
 
 - Improved button reference slot selection algorithm
 - Fixed a minor issue with policy saving when first enabling buttons on a new inventory type
+- Fixed buttons disappearing on screen resize on Fabric
+- Fixed buttons disappearing on creative inventory tab switch
 - Disabled operations in some inventories from the following mods by default:
   - Traveler's Backpack
   - Tom's Simple Storage Mod